// src/app/main/models/note.ts
export class Note {
    id: number;
    text: string;
    title: string;
    isSelected: boolean;
    createdDate: string;
    
    // constructor(_text: string, _title: string, _isSelected: boolean,_createdDate: string) {
    //     this.text = _text;
    //     this.title = _title;
    //     this.isSelected = _isSelected;
    //     this.createdDate = _createdDate;
    // }
    
}